libs-scala/resources: Factor out the test helpers into test/lib.
